requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
Django==5.0.3
django-phonenumber-field==7.3.0
Faker==24.2.0
fastapi==0.110.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
phonenumbers==8.13.32
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.10.0
yarl==1.9.4
